# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProЗапас API",
    "version": "1.0.0",
    "description": "HTTP-интерфейс между десктопным приложением и сервером.\n\nСистема обслуживает несколько складов одной компании. Каждый пользователь привязан к одному складу, и `warehouse_id` берётся из токена — в запросах он не передаётся. Сервер сам подставляет фильтр видимости, поэтому чужие заказы недостижимы.\n\nЗаказ — перемещение товара между двумя складами. Направление не хранится: для склада-заказчика заказ исходящий, для отправителя входящий.\n\nПодробности и обоснования — docs/api.md."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/orders": {
      "get": {
        "tags": [
          "Заказы"
        ],
        "summary": "Список заказов",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "total",
                    "limit",
                    "offset"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/OrderListItem"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 128
                    },
                    "limit": {
                      "type": "integer",
                      "example": 50
                    },
                    "offset": {
                      "type": "integer",
                      "example": 0
                    }
                  },
                  "description": "Страница заказов"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          }
        },
        "description": "Видны только заказы своего склада: from_warehouse_id = мой OR to_warehouse_id = мой. Подставляется сервером, задать нельзя.",
        "parameters": [
          {
            "name": "tab",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "outgoing",
                "incoming"
              ]
            },
            "description": "outgoing — мы заказали и принимаем; incoming — у нас заказали и мы отгружаем"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "created",
                  "processing",
                  "shipped",
                  "received",
                  "declined",
                  "cancelled"
                ]
              }
            },
            "description": "Можно несколько"
          },
          {
            "name": "warehouse_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Второй склад"
          },
          {
            "name": "responsible_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "created_from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "created_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Префикс номера заказа"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ]
      },
      "post": {
        "tags": [
          "Заказы"
        ],
        "summary": "Создать заказ",
        "responses": {
          "201": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          }
        },
        "description": "Склад-заказчик — наш, из токена. Наличие на чужом складе не проверяется: отправитель сам решит, сможет ли собрать.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Повтор с тем же ключом не создаёт второй заказ."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreate"
              }
            }
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "tags": [
          "Заказы"
        ],
        "summary": "Карточка заказа",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            },
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Версия документа для If-Match"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/orders/{id}/history": {
      "get": {
        "tags": [
          "Заказы"
        ],
        "summary": "История статусов",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StatusEvent"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/orders/{id}/accept": {
      "post": {
        "tags": [
          "Заказы"
        ],
        "summary": "Принять заказ",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          }
        },
        "description": "Только склад-отправитель, только в created. Одной транзакцией: статус processing, резерв остатка, пустая отгрузка.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Версия документа из ETag. Без заголовка — 428, с устаревшим — 409."
          }
        ]
      }
    },
    "/orders/{id}/decline": {
      "post": {
        "tags": [
          "Заказы"
        ],
        "summary": "Отклонить заказ",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          }
        },
        "description": "Только склад-отправитель, только в created.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Версия документа из ETag. Без заголовка — 428, с устаревшим — 409."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonBody"
              }
            }
          }
        }
      }
    },
    "/orders/{id}/cancel": {
      "post": {
        "tags": [
          "Заказы"
        ],
        "summary": "Отменить заказ",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          }
        },
        "description": "Только склад-заказчик и только пока у отгрузки нет ни одной коробки. Резерв, если был, снимается.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Версия документа из ETag. Без заголовка — 428, с устаревшим — 409."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonBody"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access-токен. Внутри — склад пользователя и права."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "description": "RFC 9457",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "example": "https://prozapas/errors/insufficient-stock"
          },
          "title": {
            "type": "string",
            "example": "Недостаточно свободного остатка"
          },
          "status": {
            "type": "integer",
            "example": 409
          },
          "detail": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
      "WarehouseBrief": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "code": {
            "type": "string",
            "example": "128"
          },
          "name": {
            "type": "string",
            "example": "Склад №1"
          },
          "owner": {
            "type": "string",
            "nullable": true,
            "example": "ООО «Сталкер Групп»"
          },
          "responsible": {
            "$ref": "#/components/schemas/UserBrief",
            "description": "Ответственный за склад. Карточка заказа показывает его для обеих сторон"
          }
        }
      },
      "UserBrief": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "example": "Кузнецов И.А."
          }
        }
      },
...
      "StatusEvent": {
        "type": "object",
        "properties": {
          "order_id": {
            "type": "integer"
          },
          "number": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "created",
              "processing",
              "shipped",
              "received",
              "declined",
              "cancelled"
            ]
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/UserBrief"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "OrderPosition": {
        "type": "object",
        "properties": {
          "article": {
            "type": "string",
            "example": "100512"
          },
          "name": {
            "type": "string"
          },
          "unit": {
            "type": "string",
            "example": "м"
          },
          "qty": {
            "type": "number",
            "format": "double",
            "description": "Дробное: в справочнике есть метры и листы"
          }
        }
      },
      "OrderListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "number": {
            "type": "string",
            "example": "2001"
          },
          "status": {
            "type": "string",
            "enum": [
              "created",
              "processing",
              "shipped",
              "received",
              "declined",
              "cancelled"
            ]
          },
          "counterparty": {
            "$ref": "#/components/schemas/WarehouseBrief",
            "description": "Второй склад: отправитель или заказчик"
          },
          "positions_count": {
            "type": "integer"
          },
          "responsible": {
            "$ref": "#/components/schemas/UserBrief"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "shipped_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "accepted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "number": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "created",
              "processing",
              "shipped",
              "received",
              "declined",
              "cancelled"
            ]
          },
          "from_warehouse": {
            "$ref": "#/components/schemas/WarehouseBrief"
          },
          "to_warehouse": {
            "$ref": "#/components/schemas/WarehouseBrief"
          },
          "responsible": {
            "$ref": "#/components/schemas/UserBrief"
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true,
            "description": "Причина отказа или отмены — одно поле на оба случая"
          },
          "positions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderPosition"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "shipped_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "accepted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "version": {
            "type": "integer"
          }
        }
      },
      "OrderCreate": {
        "type": "object",
        "required": [
          "from_warehouse_id",
          "positions"
        ],
        "properties": {
          "from_warehouse_id": {
            "type": "integer",
            "description": "Склад-отправитель. Заказчик — наш, из токена"
          },
          "comment": {
            "type": "string"
          },
          "positions": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "object",
              "required": [
                "article",
                "qty"
              ],
              "properties": {
                "article": {
                  "type": "string"
                },
                "qty": {
                  "type": "number",
                  "format": "double",
                  "description": "Дробное: в справочнике есть метры и листы"
                }
              }
            }
          }
        }
      },
      "ReasonBody": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Тело или параметры не разобрались",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Нет токена, истёк или отозван",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Нет прав на раздел или запрещено правилом",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "Объекта нет, удалён или принадлежит чужим складам",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Conflict": {
        "description": "Конфликт версий, недопустимый переход, нехватка остатка",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unprocessable": {
        "description": "Нарушено доменное правило",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "PreconditionRequired": {
        "description": "Изменение без If-Match",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
    }
  }
}